py-music-bot/requirements.txt
asyncio
PyNaCl>=1.4.0
discord.py[voice]>=1.2.5
pip
youtube_dl
toml